BufferOrch/BufferMgrDyn: do not block the m_toSync if the head needs retry
In case the head element in m_toSync needs to retry, the doTask will move to the
following items and call process<TableName> for each of the items in the m_toSync.
However, as the process<TableName> always handle the first element in m_toSync,
it results in the m_toSync being blocked if the head needs retry.

Solution: pass the element to which doTask move to process<TableName> so that
it's able to handle the following element in m_toSync

Signed-off-by: Stephen Sun <stephens@nvidia.com>