Fix dynamic buffer bug occuring in rare condition
Bug:
The buffermgrd can keep adding suffix to the buffer pool reference if
the buffer pool isn't found when it is being referenced. In most of the
cases, it's caused by wrong configuration.
Cause:
In handleBufferProfileTable, the value of each field is designated by
a (lvalue) reference to the object in the tuple, which means the object
in the tuple will be changed if the value is changed in the function.
Fix:
Pass the value of each field by value instead of reference.

Signed-off-by: Stephen Sun <stephens@nvidia.com>