Support dynamic buffer calculation on top of port auto negotiation
One logic was to take the configured speed into account when calculating the headroom
because it's the maximum possible speed at which a port runs.
As port auto negotiation supported, this logic needs to be updated accordingly:
The maximum possible speed is no longer the configured one with auto negotiation enabled.
So the speed taken into account should be:
- if port auto negotiation enabled:
  maximum speed in adv_speeds if it is configured
  maximum speed in supported_speeds otherwise
- otherwise: the configured speed

Signed-off-by: Stephen Sun <stephens@nvidia.com>