Don't call SAI API for BUFFER_POOL/PROFILE handling in case the op is DEL and the SAI OID is NULL
There is an optimization on orch: the SET event will be replaced by a DEL event
if it is still pending in m_toSync when the DEL event is coming.
This is reasonable but can cause the SAI OID be NULL in rare case:
 - The application creates an object and then destroy it after a very short period
 - The create notification is eliminated and replaced by destroy notification
 - The SAI object hasn't been created and is NULL when it is removed

This causes SAI error which eventually makes orchagent exit.
We need to avoid it in 202106 and above
(In 202012, the orchagent won't exit so it's ok to ignore the error)
The solution is not to call SAI removing interface in case SAI OID is NULL

It can happen if a user configures something which causes the accumulative headroom exceeds the limit.
In this case, the buffer profile was created and then removed in a short time.

Signed-off-by: Stephen Sun <stephens@nvidia.com>