Yet another optimization for fast reboot
Don't apply normal profiles to APPL_DB for admin down ports.
Theoretically, this should not affect SAI performance since update items
with the same key will be merged if the notifications are in m_toSync.
However, if system is under heavy workload, the notifications for
applying normal profiles can be handled before those for zero profiles
coming. In that case, it's unable to merge.

Signed-off-by: Stephen Sun <stephens@nvidia.com>