Fix maximum headroom check failure
The accumulative headroom on a port is compared with the maximum headroom supported on the port
whenever a buffer priority group is created/updated.
This depends on the maximum headroom being exposed to the STATE_DB during orchagent initialization.
However, in the cold reboot, orchagent starts slow which prevents the threshold from being exposed on time.
In this case, the buffer manager is not able to perform the headroom check
and the buffer orchagent should handle the possible failure from SAI in case the accumulative headroom exceeds the threshold.

Signed-off-by: Stephen Sun <stephens@nvidia.com>