Explicitly setting the decap TTL mode for VxLAN tunnels (#3963)
What I did
Set the decap TTL mode of VxLAN tunnels according to the ttl_mode attribute.
ttl_mode is a new attribute that can be set for each VxLAN in the VXLAN_TUNNEL table in CONFIG DB. It can be set to uniform or pipe (similar to ordinary IP-in-IP tunnels). For each VxLAN tunnel,

If ttl_mode is set, then the corresponding SAI attribute is used to set the decap TTL mode for that VxLAN tunnel on the ASIC.
Otherwise, no SAI attribute is added for the decap TTL mode (i.e., we fall back to the default ASIC behavior).
Why I did it
For the decap scenario, the default expectation is to keep the inner TTL (i.e., the pipe model). I added support for the ttl_mode attribute to be able to control this behavior if needed.