[Fixbug]: Fix delete default HFT configuration issue (#4025)
What I did
This PR is for fixing a bug that telemetry_tam object may not be deleted after the default HFT profile removed.

Why I did it
To the default profile, its group is empty, so the original checker isEmpty will be unable to capture the group deletion message.